financial-analysis-agent/pyproject.toml
[project]
name = "financial-analysis-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "yfinance>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "streamlit>=1.46.1",
    "google-genai>=1.24.0",
]
